+++ transport/common/common.go
@@ -21,5 +21,16 @@ func GetBucketAndKey(u *url.URL) (bucket, key string) {
 
 	key = strings.TrimLeft(key, "/")
 
+	// We replaced all `%` with `%25` in `imagedata.BuildImageRequest` to prevent parsing errors.
+	// Also, we replaced all `#` with `%23` to prevent cutting off the fragment part.
+	// We need to revert these replacements.
+	//
+	// It's important to revert %23 first because the original URL may also contain %23,
+	// and we don't want to mix them up.
+	bucket = strings.ReplaceAll(bucket, "%23", "#")
+	bucket = strings.ReplaceAll(bucket, "%25", "%")
+	key = strings.ReplaceAll(key, "%23", "#")
+	key = strings.ReplaceAll(key, "%25", "%")
+
 	return
 }